Fixed bug with repo file size limit overflow.

+++ FEBioStudio/RepoConnectionHandler.h
@@ -57,7 +57,7 @@ public:
 
 	QString getUsername();
 	int getUploadPermission();
-	int getSizeLimit();
+	qint64 getSizeLimit();
 	bool isAuthenticated();
 
 	void setUploadReady(bool ready);
